Suppression de la virgule dans index.js / Ajout du type pour l'input / Ajout du componentdidupdate + refactorisation de submitNumber
