+++ src/index.js
@@ -19,7 +19,6 @@ ReactDOM.render(
     <BrowserRouter>
       <App />
     </BrowserRouter>
-    ,
   </Provider>,
   document.getElementById("root")
 );
